Improving Form Structure Dissect.
1. Separators added to submenu "Change Type".
New MenuItem structure:
Change Type
  Byte
  2 Byte
  4 Byte
  8 Byte
  ----------
  Byte (Hex)
  2 Byte (Hex)
  4 Byte (Hex)
  8 Byte (Hex)
  ----------
  Float
  Double
  ----------
  String
  Unicode String
  ----------
  Array Of Byte
  Pointer

2. New MenuItem "8 Byte" added to submenu "Change Type".

3. Fixed a bug in which the Pointer(String, Unicode String) retains Hex Flag when changing the address type through the MenuItem "Change Type".

4. New MenuItem "Change all values in row".
Now you can change all values in a row with only one command.

5. New MenuItem "Add all in row to address list".
Now you can add all addresses in a row to the Address List with only one command.

6. New MenuItem "Find out what accesses this address".
The MenuItem works in the same way as in the Main Form.

7. New MenuItem "Find out what writes to this address".
The MenuItem works in the same way as in the Main Form.