Expanded group scan pointer fields to support static/dynamic/any pointers.
This commit adds additional syntax to the groupscan pointer fields so that
P:S and P:D can be used in addition to P:*. Pointers are considered static
if they point to MEM_MAPPED or MEM_IMAGE regions, and dynamic if they do
not point to such a region.

Internally, this splits the pointer lookup into two, so that there is a
separate lookup for each type (dynamic/static). The lookups are only
populated when they are used by the groupscan.