Fix lookup init bug in groupscan parser and add executable pointer scan.
There was an errant 'break;' line in the groupscan parser that caused one of the lookup tables to not be properly populated when the first pointer field used S or D pointer types. This has been fixed.

This commit also adds support for scanning for executable pointers in groupscan, using P:E syntax. This finds any pointers that point to memory that is executable, no matter whether it is static or not. Useful for finding code pointers in structs, especially in unpacked code and JIT buffers.