LUA: Change byteTableTo* functions to take optional start offset
This is so that you can have a bigger byte table and read a value that is not from the start.

This can be done purely in Lua by creating new tables that copy bytes from the original byte table
but that seems like a very cumbersome way to do it, instead changing the existing functions to take
an optional 'start offset' parameter seems like a much cleaner way to do it.