Add monoAA_FINDMONOMETHODWITHPARAMS
When a method is overridden with different parameters, this method
will let you hook the right one by specifying the parameter types.

I've run into this issue several times where the method I want to
hook has several versions.  An example is the free game on Steam
'Underminer'.   There are three versions of
`Inventory.TryRemoveItem`, this will hook the one that is called
when you use a missile, so you can just `mov al, 1; ret` to have
infinite items.   The default FINDMONOMETHOD seems to only find
the first instance of the method which might not be the one you
want to hook.

    FINDMONOMETHODWITHPARAMS(address, :Inventory:TryRemoveItem, Item,int)